refactor(admin): split large admin sections
