+++ frontend/src/admin/sections/StatsSection.svelte
@@ -11,7 +11,6 @@
   } from "$components/ui/icons.js";
   import { getContext, onMount } from "svelte";
 
-  import { fmtTrafficBytes } from "../../lib/admin/format.js";
   import Badge from "$components/ui/badge.svelte";
   import { ScrollArea } from "$components/ui/index.js";
   import * as Card from "$components/ui/card/index.js";
@@ -33,6 +32,22 @@
     inclusiveDaySpan,
     sliceLastDays,
   } from "../../lib/admin/revenueSeriesAgg.js";
+  import {
+    computeRevenueKpis,
+    formatTrafficGbCell,
+    growthBadgeVariant,
+    parsePanelBandwidth,
+    parsePanelNodeTraffic,
+    parsePanelSystem,
+    paymentDescriptionDisplay,
+    type AdminStats,
+    type CustomRangeApply,
+    type PanelNodeTraffic,
+    type PanelStats,
+    type PanelSystemMetrics,
+    type RevenueKpis,
+    type RevenuePoint,
+  } from "$lib/admin/statsDerivations";
   import type { PaymentOut, PaymentsStore } from "$lib/admin/stores/paymentsStore";
   import type { StatsState, StatsStore } from "$lib/admin/stores/statsStore";
 
@@ -40,88 +55,9 @@
   type FormatterFn = (value: unknown, currency?: string) => string;
   type DateFormatterFn = (value: unknown) => string;
   type AdminBadgeVariant = "success" | "danger" | "warning" | "muted";
-  type BadgeVariant = "default" | "outline" | "destructive" | "success" | "muted";
   type RevenueGranularity = "day" | "week" | "month";
   type RevenueRangeMode = "preset" | "custom";
-  type RevenuePoint = { date: string; amount: number };
   type IsoRange = { from: string; to: string };
-  type DynamicRecord = Record<string, unknown>;
-  type PanelStats = DynamicRecord & {
-    error?: unknown;
-    system?: DynamicRecord & {
-      users?: DynamicRecord & { statusCounts?: DynamicRecord; totalUsers?: unknown };
-      onlineStats?: DynamicRecord & { onlineNow?: unknown };
-      memory?: DynamicRecord & { total?: unknown; used?: unknown };
-      cpu?: DynamicRecord & {
-        usage?: unknown;
-        usedPercent?: unknown;
-        percent?: unknown;
-      };
-      cpuUsage?: unknown;
-      cpuLoad?: unknown;
-    };
-    bandwidth?: DynamicRecord & {
-      bandwidthLastSevenDays?: DynamicRecord & { current?: unknown };
-      bandwidthLast30Days?: DynamicRecord & { current?: unknown };
-      bandwidthLastThirtyDays?: DynamicRecord & { current?: unknown };
-    };
-    nodes?: unknown;
-    nodes_bandwidth?: DynamicRecord & { topNodes?: unknown; series?: unknown };
-  };
-  type PanelMetricRow = DynamicRecord & {
-    label: string;
-    value: string;
-    sort: number;
-    uuid: string | null;
-    online: number | null;
-  };
-  type PanelNodeTraffic = { seven: PanelMetricRow[] };
-  type PanelSystemMetrics = {
-    onlineNow: unknown;
-    active: unknown;
-    disabled: unknown;
-    expired: unknown;
-    limited: unknown;
-    totalPanelUsers: unknown;
-    memPct: number | null;
-    cpuPct: number | null;
-  };
-  type AdminStats = NonNullable<StatsState["stats"]> & {
-    financial: DynamicRecord & {
-      daily_series?: RevenuePoint[];
-      today_payments_count?: unknown;
-      today_revenue?: unknown;
-      week_revenue?: unknown;
-      month_revenue?: unknown;
-      all_time_revenue?: unknown;
-    };
-    panel: PanelStats | null;
-    recent_payments: PaymentOut[];
-    users: DynamicRecord & {
-      total_users?: number;
-      active_today?: number;
-      banned_users?: number;
-      referral_users?: number;
-      active_subscriptions?: number;
-      paid_subscriptions?: number;
-      free_subscription_users?: number;
-      trial_users?: number;
-      inactive_users?: number;
-      expired_subscription_users?: number;
-    };
-  };
-  type RevenueKpis = {
-    last7: number;
-    prev7: number;
-    growthPct: number | null;
-    avgToday: number | null;
-    total14: number;
-    maxY: number;
-    amounts: number[];
-    n: number;
-  };
-  type CustomRangeApply = { fromIso: string; toIso: string };
-  type NodeLookup = { byUuid: Map<string, number>; byName: Map<string, number> };
 
   export let at: TranslateFn;
   export let fmtDate: DateFormatterFn = (value) => String(value ?? "");
@@ -252,424 +188,6 @@
     at("date", {}, "Дата"),
   ];
   $: recentPayments = (stats?.recent_payments || []).slice(0, 10);
-
-  function isRecord(value: unknown): value is DynamicRecord {
-    return Boolean(value && typeof value === "object" && !Array.isArray(value));
-  }
-
-  function recordRows(value: unknown): DynamicRecord[] {
-    return Array.isArray(value) ? value.filter(isRecord) : [];
-  }
-
-  function formatTrafficGbCell(v: number | string | null | undefined): string {
-    if (v == null || v === "") return "—";
-    const n = Number(v);
-    if (Number.isNaN(n)) return "—";
-    let s;
-    if (Math.abs(n - Math.round(n)) < 1e-9) {
-      s = String(Math.round(n));
-    } else {
-      s = String(Math.round(n * 100) / 100);
-    }
-    return `${s} GB`;
-  }
-
-  function formatGbAmountPlain(v: number | string | null | undefined): string {
-    if (v == null || v === "") return "";
-    const n = Number(v);
-    if (Number.isNaN(n)) return "";
-    if (Math.abs(n - Math.round(n)) < 1e-9) return String(Math.round(n));
-    return String(Math.round(n * 100) / 100);
-  }
-
-  function paymentDescriptionDisplay(p: PaymentOut): string {
-    const r = p.traffic_regular_gb;
-    const pr = p.traffic_premium_gb;
-    if (r != null && pr == null) {
-      const gb = formatGbAmountPlain(r);
-      return at(
-        "payments_desc_traffic_package_regular",
-        { gb },
-        `Пакет трафика ${gb} ГБ (обычный)`
-      );
-    }
-    if (pr != null && r == null) {
-      const gb = formatGbAmountPlain(pr);
-      return at(
-        "payments_desc_traffic_package_premium",
-        { gb },
-        `Пакет трафика ${gb} ГБ (премиум)`
-      );
-    }
-    const raw = p.description && String(p.description).trim();
-    return raw || "—";
-  }
-
-  function parsePanelSystem(panel: PanelStats): PanelSystemMetrics | null {
-    const system = panel?.system;
-    if (!system || typeof system !== "object") return null;
-    const u = system.users || {};
-    const statusCounts = u.statusCounts || {};
-    const onlineStats = system.onlineStats || {};
-    const mem = system.memory || {};
-    const memTotal = Number(mem.total) || 0;
-    const memUsed = Number(mem.used) || 0;
-    const memPct = memTotal > 0 ? (memUsed / memTotal) * 100 : null;
-    const cpuRaw =
-      system.cpu?.usage ??
-      system.cpu?.usedPercent ??
-      system.cpu?.percent ??
-      system.cpuUsage ??
-      system.cpuLoad;
-    const cpuPct = Number(cpuRaw);
-    return {
-      onlineNow: onlineStats.onlineNow ?? 0,
-      active: statusCounts.ACTIVE ?? 0,
-      disabled: statusCounts.DISABLED ?? 0,
-      expired: statusCounts.EXPIRED ?? 0,
-      limited: statusCounts.LIMITED ?? 0,
-      totalPanelUsers: u.totalUsers ?? 0,
-      memPct,
-      cpuPct: Number.isFinite(cpuPct) ? cpuPct : null,
-    };
-  }
-
-  function parsePanelBandwidth(panel: PanelStats): { week: unknown; month: unknown } | null {
-    const bw = panel?.bandwidth;
-    if (!bw || typeof bw !== "object") return null;
-    const week = bw.bandwidthLastSevenDays?.current;
-    const month = bw.bandwidthLast30Days?.current ?? bw.bandwidthLastThirtyDays?.current;
-    if (week == null && month == null) return null;
-    return { week, month };
-  }
-
-  function panelRowBytes(row: DynamicRecord | null | undefined): number {
-    if (!row) return 0;
-    const total = Number(row.total);
-    if (Number.isFinite(total) && total > 0) return total;
-    const up = Number(row.uploadBytes ?? row.uplinkBytes ?? row.uplink ?? row.up ?? row.upload);
-    const down = Number(
-      row.downloadBytes ?? row.downlinkBytes ?? row.downlink ?? row.down ?? row.download
-    );
-    const sum = (Number.isFinite(up) ? up : 0) + (Number.isFinite(down) ? down : 0);
-    return sum > 0 ? sum : 0;
-  }
-
-  /** Remnawave node metrics: inboundsStats / outboundsStats use uplink+downlink per tag. */
-  function sumDirectionPair(item: DynamicRecord | null | undefined): number {
-    if (!item) return 0;
-    const combined = Number(item.total ?? item.bytes ?? item.value);
-    if (Number.isFinite(combined) && combined > 0) return combined;
-    const up = Number(
-      item.uplink ?? item.upload ?? item.uploadBytes ?? item.up ?? item.tx ?? item.sent
-    );
-    const down = Number(
-      item.downlink ?? item.download ?? item.downloadBytes ?? item.down ?? item.rx ?? item.received
-    );
-    return (Number.isFinite(up) ? up : 0) + (Number.isFinite(down) ? down : 0);
-  }
-
-  function sumTaggedStatsList(arr: unknown): number {
-    return recordRows(arr).reduce((acc, item) => acc + sumDirectionPair(item), 0);
-  }
-
-  /** Traffic bytes for one node record from GET /system/stats/nodes (current panel shape). */
-  function trafficBytesFromNodeRecord(node: DynamicRecord | null | undefined): number {
-    if (!node) return 0;
-    let b =
-      sumTaggedStatsList(node.inboundsStats) +
-      sumTaggedStatsList(node.outboundsStats) +
-      sumTaggedStatsList(node.inbounds_stats) +
-      sumTaggedStatsList(node.outbounds_stats);
-    if (b <= 0) b = panelRowBytes(node);
-    const life = Number(
-      node.totalBytesLifetime ?? node.totalBytes ?? node.bytesLifetime ?? node.totalTrafficBytes
-    );
-    if (b <= 0 && Number.isFinite(life) && life > 0) b = life;
-    return b;
-  }
-
-  function isNodeMetricsShape(row: DynamicRecord | null | undefined): boolean {
-    if (!row) return false;
-    return (
-      Array.isArray(row.inboundsStats) ||
-      Array.isArray(row.outboundsStats) ||
-      Array.isArray(row.inbounds_stats) ||
-      Array.isArray(row.outbounds_stats)
-    );
-  }
-
-  function panelRowLabel(row: DynamicRecord | null | undefined): string {
-    if (!row) return "—";
-    for (const k of ["nodeName", "node_name", "name", "nodeRemark", "remark", "label", "title"]) {
-      const v = row[k];
-      if (v != null && String(v).trim()) return String(v).trim();
-    }
-    const u = row.nodeUuid ?? row.node_uuid ?? row.uuid;
-    if (u) return `${String(u).slice(0, 8)}…`;
-    return "—";
-  }
-
-  function nodeRecordUuid(row: DynamicRecord | null | undefined): string {
-    if (!row) return "";
-    const u = row.nodeUuid ?? row.node_uuid ?? row.uuid ?? row.id;
-    return u != null ? String(u) : "";
-  }
-
-  function nodeRecordDisplayName(row: DynamicRecord | null | undefined): string {
-    if (!row) return "";
-    for (const k of ["nodeName", "node_name", "name", "label", "title", "hostname"]) {
-      const v = row[k];
-      if (v != null && String(v).trim()) return String(v).trim();
-    }
-    return "";
-  }
-
-  function nodeRecordUsersOnline(row: DynamicRecord | null | undefined): number | null {
-    if (!row) return null;
-    const raw =
-      row.usersOnline ??
-      row.users_online ??
-      row.onlineUsers ??
-      row.online_users ??
-      row.onlineUserCount ??
-      row.online_user_count ??
-      row.connectedUsers ??
-      row.connected_users ??
-      row.onlineNow;
-    const n = Number(raw);
-    if (Number.isFinite(n)) return n;
-    const mg = row.metricGroups;
-    if (isRecord(mg)) {
-      const v = Number(mg.onlineUsers ?? mg.online_users);
-      if (Number.isFinite(v)) return v;
-    }
-    return null;
-  }
-
-  /** Node list shapes from GET /system/stats/nodes (varies by panel version). */
-  function extractPanelNodesList(raw: unknown): DynamicRecord[] {
-    if (!raw) return [];
-    if (Array.isArray(raw)) return recordRows(raw);
-    if (!isRecord(raw)) return [];
-    if (Array.isArray(raw.nodes)) return recordRows(raw.nodes);
-    if (Array.isArray(raw.items)) return recordRows(raw.items);
-    if (Array.isArray(raw.data)) return recordRows(raw.data);
-    if (Array.isArray(raw.response)) return recordRows(raw.response);
-    return [];
-  }
-
-  /** UUID + display name -> online count from panel node metrics. */
-  function buildNodeOnlineLookup(panel: PanelStats): NodeLookup {
-    const byUuid = new Map<string, number>();
-    const byName = new Map<string, number>();
-    const list = extractPanelNodesList(panel?.nodes);
-    for (const node of list) {
-      if (!node || typeof node !== "object") continue;
-      const online = nodeRecordUsersOnline(node);
-      if (online == null) continue;
-      const id = nodeRecordUuid(node);
-      if (id) byUuid.set(id.toLowerCase(), online);
-      const nm = nodeRecordDisplayName(node);
-      if (nm) byName.set(nm.toLowerCase(), online);
-    }
-    return { byUuid, byName };
-  }
-
-  function formatTrafficCell(
-    bytes: number,
-    row: DynamicRecord | null | undefined,
-    stringHint: unknown
-  ): string {
-    if (bytes > 0) return fmtTrafficBytes(bytes);
-    const cur = row?.current;
-    if (typeof cur === "string" && cur.trim()) return cur.trim();
-    if (typeof stringHint === "string" && stringHint.trim()) return stringHint.trim();
-    if (isNodeMetricsShape(row) && !bytes) return fmtTrafficBytes(0);
-    return "—";
-  }
-
-  function buildNodeMetricsRows(nodes: DynamicRecord[]): PanelMetricRow[] {
-    return nodes
-      .filter((n) => n && typeof n === "object")
-      .map((node) => {
-        const bytes = trafficBytesFromNodeRecord(node);
-        const uid = nodeRecordUuid(node);
-        return {
-          label: panelRowLabel(node),
-          value: formatTrafficCell(bytes, node, ""),
-          sort: bytes,
-          uuid: uid || null,
-          online: nodeRecordUsersOnline(node),
-        };
-      })
-      .sort((a, b) => b.sort - a.sort);
-  }
-
-  /** Aggregate legacy arrays (daily rows per node, etc.). */
-  function aggregatePanelNodeRows(rows: unknown): PanelMetricRow[] {
-    const sourceRows = recordRows(rows);
-    if (!sourceRows.length) return [];
-    const map = new Map<string, { label: string; bytes: number; stringHint: string }>();
-    for (const row of sourceRows) {
-      const key = String(
-        row.nodeUuid ?? row.node_uuid ?? row.uuid ?? row.nodeName ?? row.name ?? panelRowLabel(row)
-      );
-      const prev = map.get(key) || { label: panelRowLabel(row), bytes: 0, stringHint: "" };
-      const add = isNodeMetricsShape(row) ? trafficBytesFromNodeRecord(row) : panelRowBytes(row);
-      prev.bytes += add;
-      const cur = row.current;
-      if (typeof cur === "string" && cur.trim()) prev.stringHint = cur.trim();
-      prev.label = panelRowLabel(row) || prev.label;
-      map.set(key, prev);
-    }
-    return [...map.values()]
-      .map((x) => ({
-        label: x.label,
-        value:
-          x.bytes > 0
-            ? fmtTrafficBytes(x.bytes)
-            : x.stringHint && String(x.stringHint).trim()
-              ? String(x.stringHint).trim()
-              : "—",
-        sort: x.bytes,
-        uuid: null,
-        online: null,
-      }))
-      .sort((a, b) => b.sort - a.sort);
-  }
-
-  function bandwidthRowUuid(n: DynamicRecord | null | undefined): string {
-    if (!n) return "";
-    const u = n.uuid ?? n.nodeUuid ?? n.node_uuid ?? n.id;
-    return u != null ? String(u) : "";
-  }
-
-  function attachNodeOnlineToRows(rows: PanelMetricRow[], lookup: NodeLookup): PanelMetricRow[] {
-    if (!Array.isArray(rows) || !lookup) return rows;
-    const { byUuid, byName } = lookup;
-    if (!byUuid.size && !byName.size) return rows;
-    return rows.map((r) => {
-      let o = r.online;
-      if (o == null && r.uuid) {
-        const hit = byUuid.get(String(r.uuid).toLowerCase());
-        if (hit != null) o = hit;
-      }
-      if (o == null && r.label && typeof r.label === "string") {
-        const hit = byName.get(r.label.trim().toLowerCase());
-        if (hit != null) o = hit;
-      }
-      if (o != null) return { ...r, online: o };
-      return r;
-    });
-  }
-
-  /** Panel analytics: GET /bandwidth-stats/nodes — totals per node for the selected range (bytes). */
-  function parseNodesBandwidthTop(panel: PanelStats): PanelNodeTraffic | null {
-    const nb = panel?.nodes_bandwidth;
-    if (!nb || typeof nb !== "object") return null;
-    const top = nb.topNodes;
-    const topRows = recordRows(top);
-    if (topRows.length) {
-      const rows = topRows.map((n) => {
-        const total = Number(n?.total ?? n?.bytes ?? 0);
-        const uuid = bandwidthRowUuid(n);
-        const label =
-          (typeof n?.name === "string" && n.name.trim()) ||
-          (typeof n?.nodeName === "string" && n.nodeName.trim()) ||
-          (uuid ? `${uuid.slice(0, 8)}…` : "—");
-        const directOn = Number(n?.usersOnline ?? n?.users_online ?? n?.onlineUsers);
-        const onlineInit = Number.isFinite(directOn) ? directOn : null;
-        return {
-          label,
-          value: total > 0 ? fmtTrafficBytes(total) : fmtTrafficBytes(0),
-          sort: total,
-          uuid: uuid || null,
-          online: onlineInit,
-        };
-      });
-      return { seven: rows.sort((a, b) => b.sort - a.sort) };
-    }
-    const series = nb.series;
-    const seriesRows = recordRows(series);
-    if (seriesRows.length) {
-      const rows = seriesRows.map((s) => {
-        const total = Number(s?.total ?? 0);
-        const uuid = bandwidthRowUuid(s);
-        const label =
-          (typeof s?.name === "string" && s.name.trim()) ||
-          (typeof s?.nodeName === "string" && s.nodeName.trim()) ||
-          (uuid ? `${uuid.slice(0, 8)}…` : "—");
-        const directOn = Number(s?.usersOnline ?? s?.users_online ?? s?.onlineUsers);
-        const onlineInit = Number.isFinite(directOn) ? directOn : null;
-        return {
-          label,
-          value: total > 0 ? fmtTrafficBytes(total) : fmtTrafficBytes(0),
-          sort: total,
-          uuid: uuid || null,
-          online: onlineInit,
-        };
-      });
-      return { seven: rows.sort((a, b) => b.sort - a.sort) };
-    }
-    return null;
-  }
-
-  function parsePanelNodeTraffic(panel: PanelStats): PanelNodeTraffic {
-    const onlineLookup = buildNodeOnlineLookup(panel);
-    const fromBw = parseNodesBandwidthTop(panel);
-    if (fromBw?.seven?.length) return { seven: attachNodeOnlineToRows(fromBw.seven, onlineLookup) };
-
-    const raw = panel?.nodes;
-    if (raw == null) return { seven: [] };
-
-    if (Array.isArray(raw)) {
-      if (raw.length && isNodeMetricsShape(raw[0])) {
-        return { seven: attachNodeOnlineToRows(buildNodeMetricsRows(raw), onlineLookup) };
-      }
-      return { seven: attachNodeOnlineToRows(aggregatePanelNodeRows(raw), onlineLookup) };
-    }
-
-    if (isRecord(raw)) {
-      if (Array.isArray(raw.nodes) && raw.nodes.length) {
-        return {
-          seven: attachNodeOnlineToRows(buildNodeMetricsRows(recordRows(raw.nodes)), onlineLookup),
-        };
-      }
-      if (Array.isArray(raw.lastSevenDays) && raw.lastSevenDays.length) {
-        return {
-          seven: attachNodeOnlineToRows(aggregatePanelNodeRows(raw.lastSevenDays), onlineLookup),
-        };
-      }
-    }
-
-    return { seven: [] };
-  }
-
-  function computeRevenueKpis(
-    financial: AdminStats["financial"],
-    series: RevenuePoint[]
-  ): RevenueKpis {
-    const amounts = series.map((p) => Number(p.amount) || 0);
-    const n = amounts.length;
-    const last7 = n ? amounts.slice(-7).reduce((a, b) => a + b, 0) : 0;
-    const prev7 = n > 7 ? amounts.slice(-14, -7).reduce((a, b) => a + b, 0) : 0;
-    let growthPct: number | null = null;
-    if (n >= 14 && prev7 > 0) growthPct = ((last7 - prev7) / prev7) * 100;
-    const tc = Number(financial.today_payments_count) || 0;
-    const tr = Number(financial.today_revenue) || 0;
-    const avgToday = tc > 0 ? tr / tc : null;
-    const tail14 = n >= 14 ? amounts.slice(-14) : amounts;
-    const total14 = tail14.reduce((a, b) => a + b, 0);
-    const maxY = amounts.length ? Math.max(...amounts, 1e-9) : 1e-9;
-    return { last7, prev7, growthPct, avgToday, total14, maxY, amounts, n };
-  }
-
-  function growthBadgeVariant(pct: number | null): BadgeVariant {
-    if (pct == null) return "outline";
-    if (pct >= 0) return "default";
-    return "destructive";
-  }
 
   onMount(() => {
     void statsStore.loadStats();
@@ -1394,7 +912,7 @@
                     </td>
                     <td data-label={at("provider", {}, "Провайдер")}>{p.provider}</td>
                     <td class="admin-cell-wrap" data-label={at("description", {}, "Описание")}
-                      >{paymentDescriptionDisplay(p)}</td
+                      >{paymentDescriptionDisplay(p, at)}</td
                     >
                     <td data-label={at("status", {}, "Статус")}>
                       <AdminBadge variant={paymentStatusVariant(p.status)}>{p.status}</AdminBadge>
