+++ frontend/src/admin/sections/UserDetailModal.css
@@ -1,0 +1,297 @@
+.admin-user-action-sheet {
+  border: 1px solid var(--admin-border-muted, rgba(255, 255, 255, 0.08));
+  border-radius: 12px;
+  margin-bottom: 14px;
+  overflow: hidden;
+  background: var(--admin-surface-1, rgba(255, 255, 255, 0.02));
+}
+.admin-user-action-sheet.is-dirty {
+  border-color: color-mix(in srgb, var(--warning, #f5b84b) 46%, var(--admin-border-muted));
+  background: color-mix(in srgb, var(--warning, #f5b84b) 7%, var(--admin-surface-1));
+}
+.admin-user-action-sheet .admin-dashboard-section-head {
+  padding: 12px 14px 10px;
+  margin: 0;
+  border-bottom: 1px solid var(--admin-border-muted, rgba(255, 255, 255, 0.06));
+}
+.admin-user-action-sheet-body {
+  padding: 12px 14px 12px;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+.admin-user-override-stack {
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+}
+.admin-user-grant-stack {
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+}
+.admin-user-tariff-stack {
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+}
+.admin-user-action-sheet-footer {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: flex-start;
+  gap: 12px;
+  padding: 4px 14px 12px;
+}
+.admin-override-card-footer {
+  flex-direction: column;
+  align-items: stretch;
+  gap: 10px;
+}
+.admin-override-card-toolbar {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-between;
+  gap: 10px 14px;
+  width: 100%;
+}
+.admin-override-card-toolbar .admin-btn {
+  flex: 0 0 auto;
+  min-height: 36px;
+  padding-left: 16px;
+  padding-right: 16px;
+}
+.admin-action-save-controls {
+  display: inline-flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 8px;
+  flex: 0 0 auto;
+}
+.admin-unsaved-hint {
+  color: var(--warning, #f5b84b);
+  font-weight: 600;
+}
+.admin-invalid-hint {
+  color: var(--danger, #ef4444);
+  font-weight: 600;
+}
+.admin-override-unlimited-label {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  font-size: 13px;
+  color: var(--admin-text, inherit);
+  cursor: pointer;
+  user-select: none;
+  min-height: 36px;
+}
+@media (max-width: 520px) {
+  .admin-override-card-toolbar {
+    flex-direction: column;
+    align-items: stretch;
+  }
+  .admin-override-card-toolbar .admin-btn {
+    width: 100%;
+  }
+  .admin-action-save-controls {
+    width: 100%;
+    align-items: stretch;
+    justify-content: space-between;
+    flex-wrap: wrap;
+  }
+  .admin-action-save-controls .admin-btn {
+    flex: 1 1 180px;
+  }
+}
+.admin-override-status-lines {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  flex: 1;
+  min-width: 0;
+  font-size: 12px;
+  line-height: 1.35;
+}
+.admin-user-dialog .admin-actions-tab {
+  padding-bottom: 14px;
+}
+.admin-user-action-sheet--regular-override {
+  margin-top: 10px;
+}
+.admin-user-action-sheet--extend {
+  margin-bottom: 0;
+}
+.admin-user-action-sheet--tariff {
+  margin-top: 10px;
+}
+.admin-user-action-sheet--traffic-grant {
+  margin-top: 10px;
+}
+.admin-user-action-sheet .admin-user-tariff-select {
+  width: 100%;
+  max-width: 100%;
+}
+.admin-user-action-sheet .admin-grant-kind-select {
+  width: 100%;
+  max-width: 100%;
+}
+.admin-avatar-preview-trigger {
+  padding: 0;
+  appearance: none;
+}
+.admin-avatar-preview-trigger img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+.admin-avatar-preview-trigger:disabled {
+  cursor: default;
+}
+.admin-avatar-preview-trigger.is-clickable {
+  cursor: zoom-in;
+}
+.admin-avatar-preview-trigger.is-clickable:hover,
+.admin-avatar-preview-trigger.is-clickable:focus-visible {
+  border-color: var(--accent);
+  box-shadow: 0 0 0 3px color-mix(in srgb, var(--accent) 18%, transparent);
+}
+.admin-user-summary-actions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  margin-top: 6px;
+}
+.admin-user-ref-row {
+  grid-template-columns: 130px minmax(0, 1fr) auto;
+  align-items: center;
+}
+.admin-user-ref-row .admin-btn {
+  flex: 0 0 auto;
+}
+.admin-user-ref-value {
+  display: grid;
+  gap: 2px;
+  min-width: 0;
+}
+.admin-user-ref-value small {
+  color: var(--admin-muted);
+  font-size: 11px;
+  font-weight: 500;
+}
+.admin-user-referrals-dialog {
+  width: min(760px, calc(100vw - 28px));
+  max-height: min(760px, calc(100dvh - 28px));
+}
+.admin-user-referrals-body {
+  display: grid;
+  gap: 12px;
+  min-width: 0;
+}
+.admin-user-referrals-table-wrap {
+  min-height: 120px;
+}
+.admin-referral-user-cell {
+  display: grid;
+  gap: 2px;
+  min-width: 0;
+}
+.admin-referral-user-cell strong {
+  color: var(--admin-text);
+  font-weight: 650;
+  word-break: break-word;
+}
+.admin-referral-user-cell small {
+  color: var(--admin-muted);
+  font-size: 12px;
+  word-break: break-word;
+}
+.admin-referral-user-actions {
+  text-align: right;
+}
+@media (max-width: 560px) {
+  .admin-user-ref-row {
+    grid-template-columns: minmax(0, 1fr) auto;
+    gap: 4px 8px;
+  }
+  .admin-user-ref-row > span {
+    grid-column: 1 / -1;
+  }
+}
+.admin-avatar-dialog {
+  display: grid;
+  grid-template-rows: auto minmax(0, 1fr);
+  width: min(920px, calc(100vw - 28px));
+  height: min(820px, calc(100dvh - 28px));
+  max-height: calc(100dvh - 28px);
+  gap: 10px;
+  padding: 12px;
+  overflow: hidden;
+}
+.admin-avatar-preview {
+  display: grid;
+  place-items: center;
+  min-height: 0;
+  width: 100%;
+  height: 100%;
+  padding: 4px;
+  overflow: hidden;
+}
+.admin-avatar-preview img {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+  border-radius: 14px;
+  border: 1px solid var(--admin-border);
+  background: var(--admin-surface-2);
+}
+@media (max-width: 640px) {
+  .dialog:has(.admin-avatar-dialog) {
+    padding: max(8px, env(safe-area-inset-top)) max(8px, env(safe-area-inset-right))
+      max(8px, env(safe-area-inset-bottom)) max(8px, env(safe-area-inset-left));
+  }
+  .admin-avatar-dialog {
+    width: calc(100vw - 16px);
+    height: min(88dvh, calc(100dvh - 16px));
+    max-height: calc(100dvh - 16px);
+    border-radius: 18px;
+    padding: 10px;
+  }
+}
+.admin-user-dialog .admin-user-logs-tab {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+.admin-user-logs-head {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+  flex-wrap: wrap;
+}
+.admin-user-logs-meta {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  font-size: 13px;
+}
+.admin-user-logs-wrap {
+  min-height: 120px;
+}
+.admin-user-log-event {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  flex-wrap: wrap;
+}
+.admin-user-log-content {
+  white-space: pre-wrap;
+  word-break: break-word;
+  max-width: 520px;
+}
+@media (max-width: 640px) {
+  .admin-user-log-content {
+    max-width: 100%;
+  }
+}
